requirements.txt
future==0.15.2
pkg-resources==0.0.0
psutil==4.2.0
python-telegram-bot==4.1.1
